+++ src/app/components/summary/summary.component.ts
@@ -7,6 +7,8 @@ import { MatTableModule } from "@angular/material/table";
 import { LoadingService } from "../../services/loading.service";
 import { TranslateModule } from "@ngx-translate/core";
 import { MatSortModule, Sort } from "@angular/material/sort";
+import { WeightedSumService } from "../../services/weighted-sum.service";
+import { TopsisService } from "../../services/topsis.service";
 
 @Component({
     selector: "app-summary",
@@ -21,39 +23,51 @@ import { MatSortModule, Sort } from "@angular/material/sort";
     templateUrl: "./summary.component.html",
 })
 export class SummaryComponent {
-    sortedData!: IAlternative[];
+    sortedDataWeightedSum!: IAlternative[];
+    sortedDataTopsis!: IAlternative[];
 
     constructor(
         private alternativeService: AlternativeService,
-        private loadingService: LoadingService
+        private loadingService: LoadingService,
+        private weightedSumService: WeightedSumService,
+        private topsisService: TopsisService
     ) {}
 
     ngOnInit() {
-        this.alternativeService.calculateWeightedSums();
-        this.sortedData = this.sortBySum(
+        this.weightedSumService.calculateWeightedSums();
+        this.sortedDataWeightedSum = this.sortByWeightedSum(
             this.alternativeService.alternatives
         ).slice();
-        console.log("a1".split("a"));
+        this.topsisService.runAllCalculations();
+        this.sortedDataTopsis = this.sortByTopsis(
+            this.alternativeService.alternatives
+        ).slice();
     }
 
     ngAfterViewInit(): void {
         this.loadingService.hide();
     }
 
-    sortBySum(alternatives: IAlternative[]) {
+    sortByWeightedSum(alternatives: IAlternative[]) {
         return alternatives.sort((a, b) => {
             return b.values.weightedSum! - a.values.weightedSum!;
+        });
+    }
+
+    sortByTopsis(alternatives: IAlternative[]) {
+        return alternatives.sort((a, b) => {
+            return b.topsisValues?.finalValue! - a.topsisValues?.finalValue!;
         });
     }
 
     sortData(sort: Sort) {
         const data = this.alternativeService.alternatives.slice();
         if (!sort.active || sort.direction === "") {
-            this.sortedData = data;
+            this.sortedDataWeightedSum = data;
             return;
         }
 
-        this.sortedData = data.sort((a, b) => {
+        this.sortedDataWeightedSum = data.sort((a, b) => {
             const isAsc = sort.direction === "asc";
             switch (sort.active) {
                 case "id":
@@ -68,6 +82,12 @@ export class SummaryComponent {
                     return this.compare(
                         a.values.weightedSum!,
                         b.values.weightedSum!,
+                        isAsc
+                    );
+                case "topsis":
+                    return this.compare(
+                        a.topsisValues!.finalValue!,
+                        b.topsisValues!.finalValue!,
                         isAsc
                     );
                 default: