+++ src/app/constants/version.constants.ts
@@ -1,2 +1,2 @@
-export const version = "alpha 0.3.2";
-export const date = "2024-03-20";
+export const version = "alpha 0.4.0";
+export const date = "2024-04-04";