+++ src/app/models/alternative.model.ts
@@ -10,7 +10,20 @@ export interface IAlternative {
         };
         normalized?: {
             [key: string]: number;
-        }
+        };
         weightedSum?: number;
-    }
+    };
+    topsisValues?: {
+        normalized?: {
+            [key: string]: number;
+        };
+        calculated?: {
+            [key: string]: number;
+        };
+        idealValue?: number;
+        basalValue?: number;
+        vPlus?: number;
+        vMinus?: number;
+        finalValue?: number;
+    };
 }